// src/types/api/resources/CapitalGrantOffersResource.ts
export interface paths {
    "/v1/capital/grantOffers/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Create Grant Offer
         * @description This action verifies offer conditions and makes grant ready for user to accept it
         */
        post: operations["createGrantOffer"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/capital/grantOffers/dynamic": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Dynamic Grant Offer
         * @description Get grant offer details for specific parameters
         */
        get: operations["getDynamicGrantOffer"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/capital/grantOffers/dynamic/configuration": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get dynamic offers
         * @description Given dynamic offers for the account holder
         */
        get: operations["getDynamicGrantOffersConfiguration"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        Amount: {
            /** @description The three-character [ISO currency code](https://docs.adyen.com/development-resources/currency-codes#currency-codes) of the amount. */
            currency: string;
            /**
             * Format: int64
             * @description The numeric value of the amount, in [minor units](https://docs.adyen.com/development-resources/currency-codes#minor-units).
             */
            value: number;
        };
        GrantOfferResponseDTO: {
            /** Format: int32 */
            aprBasisPoints?: number;
            /** Format: int32 */
            expectedRepaymentPeriodDays: number;
            feesAmount: components["schemas"]["Amount"];
            grantAmount: components["schemas"]["Amount"];
            id: string;
            /** Format: int32 */
            maximumRepaymentPeriodDays?: number;
            /** Format: int32 */
            repaymentRate: number;
            thresholdAmount: components["schemas"]["Amount"];
            totalAmount: components["schemas"]["Amount"];
        };
        CreateGrantOfferRequestDTO: {
            /** Format: int64 */
            amount: number;
            currency: string;
        };
        DynamicOffersResponseDTO: {
            maxAmount: components["schemas"]["Amount"];
            minAmount: components["schemas"]["Amount"];
            /** Format: int32 */
            step: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    createGrantOffer: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["CreateGrantOfferRequestDTO"];
            };
        };
        responses: {
            /** @description OK - the request has succeeded. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GrantOfferResponseDTO"];
                };
            };
        };
    };
    getDynamicGrantOffer: {
        parameters: {
            query: {
                amount: number;
                currency: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK - the request has succeeded. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GrantOfferResponseDTO"];
                };
            };
        };
    };
    getDynamicGrantOffersConfiguration: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK - the request has succeeded. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["DynamicOffersResponseDTO"];
                };
            };
        };
    };
}