// src/types/api/resources/CapitalMissingActionsResource.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/v1/capital/grants/missingActions/anaCredit": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get AnaCredit action details
         * @description Add @Operation annotation to provide a description
         */
        get: operations["anaCreditActionDetails"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/capital/grants/missingActions/signToS": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get SignToS action details
         * @description Add @Operation annotation to provide a description
         */
        get: operations["signToSActionDetails"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        MissingAnaCreditActionDetailsResponseDTO: {
            url: string;
        };
        MissingTOSActionDetailsResponseDTO: {
            url: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    anaCreditActionDetails: {
        parameters: {
            query: {
                redirectUrl: string;
                locale: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK - the request has succeeded. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["MissingAnaCreditActionDetailsResponseDTO"];
                };
            };
        };
    };
    signToSActionDetails: {
        parameters: {
            query: {
                redirectUrl: string;
                locale: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK - the request has succeeded. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["MissingTOSActionDetailsResponseDTO"];
                };
            };
        };
    };
}